+++ R/checkArgumentsPDTMT.R
@@ -17,8 +17,8 @@
     subtractBaseline, baselineGroup, normMethod, spikeFeatures, stattest,
     minNbrValidValues, minlFC, samSignificance, nperm, volcanoAdjPvalThr,
     volcanoLog2FCThr, volcanoMaxFeatures, volcanoS0, volcanoFeaturesToLabel,
-    addInteractiveVolcanos, interactiveDisplayColumns, complexFDRThr,
-    maxNbrComplexesToPlot, seed,
+    addInteractiveVolcanos, interactiveDisplayColumns, interactiveGroupColumn,
+    complexFDRThr, maxNbrComplexesToPlot, seed,
     includeFeatureCollections, minSizeToKeepSet, customComplexes,
     complexSpecies, complexDbPath, stringVersion, stringDir, linkTableColumns,
     customYml, doRender, generateQCPlot
@@ -181,6 +181,7 @@
     .assertScalar(x = allPairwiseComparisons, type = "logical")
     .assertScalar(x = addInteractiveVolcanos, type = "logical")
     .assertVector(x = interactiveDisplayColumns, type = "character", allowNULL = TRUE)
+    .assertScalar(x = interactiveGroupColumn, type = "character", allowNULL = TRUE)
     .assertScalar(x = singleFit, type = "logical")
     .assertScalar(x = subtractBaseline, type = "logical")
     .assertScalar(x = baselineGroup, type = "character")