+++ R/plotVolcano.R
@@ -297,6 +297,10 @@
 #' @param interactiveDisplayColumns Character vector (or \code{NULL})
 #'     indicating which columns of \code{res} to include in the tooltip for the
 #'     interactive volcano plots. The default shows the feature ID.
+#' @param interactiveGroupColumn Character scalar (or \code{NULL}, default)
+#'     indicating the column to group points by in the interactive volcano
+#'     plot. Hovering over a point will highlight all other points with the
+#'     same value of this column.
 #' @param maxTextWidthBarplot Numeric scalar giving the maximum allowed width
 #'     for text labels in the bar plot of log-fold changes. If not \code{NULL},
 #'     the size of the labels will be scaled down in an attempt to keep the
@@ -311,7 +315,7 @@
 #' @export
 #' @importFrom ggplot2 geom_point aes labs
 #' @importFrom ggrepel geom_text_repel
-#' @importFrom ggiraph geom_point_interactive
+#' @importFrom ggiraph geom_point_interactive opts_hover girafe_options
 #' @importFrom dplyr filter arrange between row_number desc
 #' @importFrom rlang .data
 #' @importFrom utils stack
@@ -328,6 +332,7 @@ plotVolcano <- function(sce, res, testType, xv = NULL, yv = NULL, xvma = NULL,
                         xlabma = "Average abundance",
                         labelOnlySignificant = TRUE,
                         interactiveDisplayColumns = NULL,
+                        interactiveGroupColumn = NULL,
                         maxTextWidthBarplot = NULL) {
 
     .assertScalar(x = baseFileName, type = "character", allowNULL = TRUE)
@@ -385,6 +390,7 @@ plotVolcano <- function(sce, res, testType, xv = NULL, yv = NULL, xvma = NULL,
     .assertScalar(x = xlabma, type = "character")
     .assertScalar(x = labelOnlySignificant, type = "logical")
     .assertVector(x = interactiveDisplayColumns, type = "character", allowNULL = TRUE)
+    .assertScalar(x = interactiveGroupColumn, type = "character", allowNULL = TRUE)
     .assertScalar(x = maxTextWidthBarplot, type = "numeric", allowNULL = TRUE)
 
     ## If the 'einprotLabel' column is not available, create it using the 'pid' column
@@ -395,6 +401,13 @@ plotVolcano <- function(sce, res, testType, xv = NULL, yv = NULL, xvma = NULL,
     ## -------------------------------------------------------------------------
     ## Add column with interactive labels
     ## -------------------------------------------------------------------------
+    if (!is.null(interactiveGroupColumn) && !(interactiveGroupColumn %in%
+                                              colnames(res))) {
+        warning("The interactive group column (", interactiveGroupColumn,
+                ") is not present in the data")
+        interactiveGroupColumn <- NULL
+    }
+
     if (is.null(interactiveDisplayColumns)) {
         interactiveDisplayColumns <- "einprotLabel"
     }
@@ -426,13 +439,17 @@ plotVolcano <- function(sce, res, testType, xv = NULL, yv = NULL, xvma = NULL,
             dplyr::pull("lab")
     }
 
+    ## -------------------------------------------------------------------------
     ## Make "base" volcano plot
+    ## -------------------------------------------------------------------------
     ggbase <- .makeBaseVolcano(res = res, testType = testType, xv = cols$xv, yv = cols$yv,
                                plotnote = plotnote, plottitle = plottitle,
                                plotsubtitle = plotsubtitle, curveparam = curveparam,
                                xlab = xlab, ylab = ylab)
 
+    ## -------------------------------------------------------------------------
     ## Add colors for significant features
+    ## -------------------------------------------------------------------------
     ggtest <- ggbase +
         ggplot2::geom_point(fill = "lightgrey", color = "grey",
                             pch = 21, size = 1.5) +
@@ -471,15 +488,33 @@ plotVolcano <- function(sce, res, testType, xv = NULL, yv = NULL, xvma = NULL,
     ## -------------------------------------------------------------------------
     ## Interactive version
     ## -------------------------------------------------------------------------
-    ggint <- ggbase +
-        ggiraph::geom_point_interactive(
-            aes(tooltip = .data$intLabel),
-            fill = "lightgrey", color = "grey",
-            pch = 21, size = 1.5) +
-        ggiraph::geom_point_interactive(
-            data = res %>% dplyr::filter(.data[[cols$volcind]]),
-            fill = "red", color = "grey",
-            pch = 21, size = 1.5)
+    if (!is.null(interactiveGroupColumn)) {
+        ggint <- ggbase +
+            ggiraph::geom_point_interactive(
+                aes(tooltip = .data$intLabel, data_id = .data[[interactiveGroupColumn]]),
+                fill = "lightgrey", color = "grey",
+                pch = 21, size = 1.5) +
+            ggiraph::geom_point_interactive(
+                data = res %>% dplyr::filter(.data[[cols$volcind]]),
+                aes(tooltip = .data$intLabel, data_id = .data[[interactiveGroupColumn]]),
+                fill = "red", color = "grey",
+                pch = 21, size = 1.5) +
+            labs(subtitle = paste0(plotsubtitle, "\nGrouped by ",
+                                   interactiveGroupColumn))
+    } else {
+        ggint <- ggbase +
+            ggiraph::geom_point_interactive(
+                aes(tooltip = .data$intLabel),
+                fill = "lightgrey", color = "grey",
+                pch = 21, size = 1.5) +
+            ggiraph::geom_point_interactive(
+                data = res %>% dplyr::filter(.data[[cols$volcind]]),
+                fill = "red", color = "grey",
+                pch = 21, size = 1.5)
+    }
+    ggint <- ggiraph::girafe(ggobj = ggint)
+    ggint <- ggiraph::girafe_options(
+        ggint, ggiraph::opts_hover(css = "fill-opacity:1;fill:blue;stroke:blue;") )
 
     ## -------------------------------------------------------------------------
     ## MA plot
@@ -626,7 +661,7 @@ plotVolcano <- function(sce, res, testType, xv = NULL, yv = NULL, xvma = NULL,
             dev.off()
         }
     }
-    return(list(gg = ggtest, ggint = ggiraph::girafe(ggobj = ggint),
+    return(list(gg = ggtest, ggint = ggint,
                 ggma = ggma, ggwf = ggwf, ggbar = ggbar,
                 pidLabelVolcano = pidLabelVolcano))
 }